feat: file re-factor internal/webrequest
