+++ internal/webrequest/process.go
@@ -18,14 +18,268 @@ package webrequest
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
+	"time"
 
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	promoterv1alpha1 "github.com/argoproj-labs/gitops-promoter/api/v1alpha1"
 	"github.com/argoproj-labs/gitops-promoter/internal/types/constants"
 	"github.com/argoproj-labs/gitops-promoter/internal/utils"
 )
+
+// ProcessWebRequestCommitStatusInput carries shared dependencies for WRCS processing
+// (per-environment context and promotionstrategy context).
+type ProcessWebRequestCommitStatusInput struct {
+	Evaluator              *Evaluator
+	HttpExec               HTTPEXecutor
+	WebRequestCommitStatus *promoterv1alpha1.WebRequestCommitStatus
+	PromotionStrategy      *promoterv1alpha1.PromotionStrategy
+	// NamespaceMeta is passed into TemplateData for template rendering.
+	NamespaceMeta         NamespaceMetadata
+	CommitEmitter         CommitStatusEmitter
+	RenderedHTTPCollector RenderedHTTPCollector // optional; when non-nil, CollectRenderedHTTP is called after a successful template render when the trigger fires
+}
+
+// ProcessWebRequestCommitStatusEnvironmentsOutput is the computed status and CommitStatus list for one reconcile.
+type ProcessWebRequestCommitStatusEnvironmentsOutput struct {
+	Environments         []promoterv1alpha1.WebRequestCommitStatusEnvironmentStatus
+	CommitStatuses       []*promoterv1alpha1.CommitStatus
+	TransitionedBranches []string
+}
+
+// ProcessWebRequestCommitStatusPromotionStrategyOutput is the computed status for promotionstrategy context.
+type ProcessWebRequestCommitStatusPromotionStrategyOutput struct {
+	PromotionStrategyContext *promoterv1alpha1.WebRequestCommitStatusPromotionStrategyContextStatus
+	CommitStatuses           []*promoterv1alpha1.CommitStatus
+	TransitionedBranches     []string
+	ApplicableEnvsEmpty      bool
+	PollingAllSuccessSkip    bool
+}
+
+// RenderedHTTPRequest is a fully rendered HTTP request template snapshot (diagnostics).
+// Branch is set per environment in environments context; empty for promotionstrategy shared request.
+type RenderedHTTPRequest struct {
+	Branch  string
+	Method  string
+	URL     string
+	Headers map[string]string
+	Body    string
+}
+
+// CommitStatusEmitter creates or updates one CommitStatus per reconcile step (SSA upsert in the controller,
+// local render in the simulator).
+type CommitStatusEmitter interface {
+	EmitCommitStatus(
+		ctx context.Context,
+		wrcs *promoterv1alpha1.WebRequestCommitStatus,
+		repositoryRefName, branch, sha string,
+		phase promoterv1alpha1.CommitStatusPhase,
+		td TemplateData,
+	) (*promoterv1alpha1.CommitStatus, error)
+}
+
+// RenderedHTTPCollector receives fully rendered HTTP templates when the trigger fires (e.g. simulator output).
+// Leave nil in production: HTTPEXecutor already renders on the wire; collecting here would duplicate template work.
+type RenderedHTTPCollector interface {
+	CollectRenderedHTTP(r RenderedHTTPRequest)
+}
+
+// evaluateTriggerDecision determines whether the HTTP request should fire this reconcile.
+// For polling mode it checks whether the polling interval has elapsed since lastRequestTime.
+// For trigger mode it evaluates the trigger expression and optionally the trigger output expression.
+func evaluateTriggerDecision(
+	ctx context.Context,
+	evaluator *Evaluator,
+	mode promoterv1alpha1.ModeSpec,
+	td TemplateData,
+	lastRequestTime *metav1.Time,
+) (triggerDecision, error) {
+	logger := log.FromContext(ctx)
+	shouldFire := true
+	var newTriggerData map[string]any
+
+	if mode.Polling != nil && lastRequestTime != nil {
+		if elapsed := time.Since(lastRequestTime.Time); elapsed < mode.Polling.Interval.Duration {
+			logger.V(4).Info("Within polling interval, skipping HTTP request",
+				"elapsed", elapsed, "interval", mode.Polling.Interval.Duration)
+			shouldFire = false
+		}
+	}
+
+	if mode.Trigger != nil {
+		sf, ntd, err := evaluator.evaluateTriggerWhenBranch(ctx, mode.Trigger, td)
+		if err != nil {
+			return triggerDecision{}, fmt.Errorf("failed to evaluate trigger.when: %w", err)
+		}
+		shouldFire = sf
+		newTriggerData = ntd
+	}
+
+	return triggerDecision{ShouldFire: shouldFire, NewTriggerData: newTriggerData}, nil
+}
+
+// validationResultFromHTTPResponse runs response extraction and success.when evaluation after
+// an HTTP response is available. It updates td.ResponseOutput when response data JSON is present.
+func validationResultFromHTTPResponse(
+	ctx context.Context,
+	evaluator *Evaluator,
+	wrcs *promoterv1alpha1.WebRequestCommitStatus,
+	td TemplateData,
+	response HTTPResponse,
+) (validationResult, error) {
+	logger := log.FromContext(ctx)
+
+	now := metav1.Now()
+	lastRequestTime := &now
+	lastResponseStatusCode := &response.StatusCode
+
+	logger.V(4).Info("HTTP response received", "statusCode", response.StatusCode)
+
+	var responseDataJSON *apiextensionsv1.JSON
+	if wrcs.Spec.Mode.Trigger != nil && wrcs.Spec.Mode.Trigger.Response != nil {
+		extractedData, err := evaluator.evaluateResponseDataExpression(ctx, wrcs.Spec.Mode.Trigger.Response.Output.Expression, response)
+		if err != nil {
+			return validationResult{}, fmt.Errorf("failed to evaluate response data expression: %w", err)
+		}
+
+		responseDataBytes, err := json.Marshal(extractedData)
+		if err != nil {
+			return validationResult{}, fmt.Errorf("failed to marshal response data: %w", err)
+		}
+		responseDataJSON = &apiextensionsv1.JSON{Raw: responseDataBytes}
+	}
+
+	if responseDataJSON != nil {
+		if data, err := unmarshalJSONMap(responseDataJSON); err == nil && data != nil {
+			td.ResponseOutput = data
+		}
+	}
+
+	exprData := successWhenExprData(td, &response)
+	exprData, err := evaluator.enrichWhenExprEnv(ctx, wrcs.Spec.Success.When, exprData)
+	if err != nil {
+		return validationResult{}, fmt.Errorf("failed to evaluate success.when.variables: %w", err)
+	}
+	phase, phasePerBranch, err := evaluateSuccessPhase(ctx, evaluator, wrcs, exprData)
+	if err != nil {
+		return validationResult{}, fmt.Errorf("failed to evaluate validation expression: %w", err)
+	}
+
+	successDataJSON, err := evaluateSuccessOutput(ctx, evaluator, wrcs, exprData)
+	if err != nil {
+		return validationResult{}, fmt.Errorf("failed to evaluate success.when.output expression: %w", err)
+	}
+
+	return validationResult{
+		Phase:                  phase,
+		PhasePerBranch:         phasePerBranch,
+		LastRequestTime:        lastRequestTime,
+		LastResponseStatusCode: lastResponseStatusCode,
+		ResponseDataJSON:       responseDataJSON,
+		SuccessDataJSON:        successDataJSON,
+	}, nil
+}
+
+// validationResultCarryForward evaluates success.when with Response=nil when the trigger does not fire,
+// carrying forward last HTTP metadata from lastState.
+func validationResultCarryForward(
+	ctx context.Context,
+	evaluator *Evaluator,
+	wrcs *promoterv1alpha1.WebRequestCommitStatus,
+	td TemplateData,
+	lastState lastReconciledState,
+) (validationResult, error) {
+	exprData := successWhenExprData(td, nil)
+	var err error
+	exprData, err = evaluator.enrichWhenExprEnv(ctx, wrcs.Spec.Success.When, exprData)
+	if err != nil {
+		return validationResult{}, fmt.Errorf("failed to evaluate success.when.variables: %w", err)
+	}
+	phase, phasePerBranch, err := evaluateSuccessPhase(ctx, evaluator, wrcs, exprData)
+	if err != nil {
+		return validationResult{}, fmt.Errorf("failed to evaluate success.when expression: %w", err)
+	}
+
+	successDataJSON, err := evaluateSuccessOutput(ctx, evaluator, wrcs, exprData)
+	if err != nil {
+		return validationResult{}, fmt.Errorf("failed to evaluate success.when.output expression: %w", err)
+	}
+
+	return validationResult{
+		Phase:                  phase,
+		PhasePerBranch:         phasePerBranch,
+		LastRequestTime:        lastState.LastRequestTime,
+		LastResponseStatusCode: lastState.LastResponseStatusCode,
+		ResponseDataJSON:       lastState.ResponseOutput,
+		SuccessDataJSON:        successDataJSON,
+	}, nil
+}
+
+// fireOrCarryForward runs the HTTP executor when decision.ShouldFire, otherwise carries forward
+// without a new HTTP response.
+func fireOrCarryForward(
+	ctx context.Context,
+	evaluator *Evaluator,
+	wrcs *promoterv1alpha1.WebRequestCommitStatus,
+	td TemplateData,
+	decision triggerDecision,
+	lastState lastReconciledState,
+	exec HTTPEXecutor,
+) (validationResult, error) {
+	if !decision.ShouldFire {
+		return validationResultCarryForward(ctx, evaluator, wrcs, td, lastState)
+	}
+	resp, err := exec.Execute(ctx, wrcs, td)
+	if err != nil {
+		return validationResult{}, fmt.Errorf("HTTP request execution: %w", err)
+	}
+	return validationResultFromHTTPResponse(ctx, evaluator, wrcs, td, resp)
+}
+
+func evaluateSuccessPhase(
+	ctx context.Context,
+	evaluator *Evaluator,
+	wrcs *promoterv1alpha1.WebRequestCommitStatus,
+	exprData map[string]any,
+) (promoterv1alpha1.CommitStatusPhase, map[string]promoterv1alpha1.CommitStatusPhase, error) {
+	if wrcs.Spec.Mode.Context == promoterv1alpha1.ContextPromotionStrategy {
+		phase, phasePerBranch, err := evaluator.evaluateValidationExpressionForPromotionStrategy(ctx, wrcs.Spec.Success.When.Expression, exprData)
+		if err != nil {
+			return phase, phasePerBranch, fmt.Errorf("failed to evaluate validation expression (promotionstrategy context): %w", err)
+		}
+		return phase, phasePerBranch, nil
+	}
+	passed, err := evaluator.evaluateValidationExpression(ctx, wrcs.Spec.Success.When.Expression, exprData)
+	if err != nil {
+		return "", nil, fmt.Errorf("failed to evaluate validation expression: %w", err)
+	}
+	if passed {
+		return promoterv1alpha1.CommitPhaseSuccess, nil, nil
+	}
+	return promoterv1alpha1.CommitPhasePending, nil, nil
+}
+
+func evaluateSuccessOutput(
+	ctx context.Context,
+	evaluator *Evaluator,
+	wrcs *promoterv1alpha1.WebRequestCommitStatus,
+	exprData map[string]any,
+) (*apiextensionsv1.JSON, error) {
+	if wrcs.Spec.Success.When.Output == nil || wrcs.Spec.Success.When.Output.Expression == "" {
+		return nil, nil
+	}
+
+	extractedData, err := evaluator.evaluateSuccessDataExpression(ctx, wrcs.Spec.Success.When.Output.Expression, exprData)
+	if err != nil {
+		return nil, fmt.Errorf("failed to evaluate success data expression: %w", err)
+	}
+
+	return marshalJSONMap(extractedData)
+}
 
 // renderHTTPRequestTemplates renders URL, headers, and body from WebRequestCommitStatus HTTP templates.
 func renderHTTPRequestTemplates(wrcs *promoterv1alpha1.WebRequestCommitStatus, td TemplateData) (RenderedHTTPRequest, error) {
