+++ internal/webrequest/types.go
@@ -21,8 +21,21 @@ limitations under the License.
 package webrequest
 
 import (
+	"context"
+	"encoding/json"
+	"errors"
+	"fmt"
+
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
 	promoterv1alpha1 "github.com/argoproj-labs/gitops-promoter/api/v1alpha1"
 )
+
+// ErrHTTPResponseRequiredWhenTriggerFires is returned by a simulator HTTPEXecutor when the trigger
+// would fire but no mock HTTPResponse was provided. Callers may wrap it with fmt.Errorf("... %w", err)
+// (for example to attach the environment branch).
+var ErrHTTPResponseRequiredWhenTriggerFires = errors.New("HTTPResponse is required when the trigger fires (fill in a mock response, or craft inputs so the trigger does not fire)")
 
 // TemplateData is the data passed to Go templates when rendering URL, headers, body, and description,
 // and (via triggerExprData / successWhenExprData) to expr expressions.
@@ -65,4 +78,115 @@ type HTTPResponse struct {
 // description/URL templates.
 type triggerResult struct {
 	Trigger bool
+}
+
+// triggerDecision is the result of evaluateTriggerDecision: whether to perform the HTTP round-trip
+// and any trigger output data for the next reconcile.
+type triggerDecision struct {
+	NewTriggerData map[string]any
+	ShouldFire     bool
+}
+
+// validationResult holds the outcome of processing a fire or carry-forward path. Phase is derived
+// from the success.when expression and is written to CommitStatus / simulated status.
+//
+// When context is promotionstrategy and the success expression returns an object { defaultPhase?, environments? },
+// PhasePerBranch is set and used to set each environment's CommitStatus phase; Phase is the default for branches not in the per-branch map.
+//
+// ResponseDataJSON is set only in trigger mode when response.output.expression is configured: it is the
+// JSON-serialized map returned by the data expression (extract/transform from the HTTP response).
+// It is stored in WebRequestCommitStatusEnvironmentStatus.ResponseOutput so it persists across
+// reconciles. On the next run it is unmarshalled into templateData.ResponseOutput, so the trigger
+// expression can read it (e.g. ResponseOutput.buildUrl). When upserting the CommitStatus it is also
+// passed into the description and URL templates as ResponseOutput, so the SCM status can show links
+// or text derived from the response (e.g. a link to the CI run).
+type validationResult struct {
+	LastRequestTime        *metav1.Time
+	LastResponseStatusCode *int
+	ResponseDataJSON       *apiextensionsv1.JSON
+	SuccessDataJSON        *apiextensionsv1.JSON
+	PhasePerBranch         map[string]promoterv1alpha1.CommitStatusPhase
+	Phase                  promoterv1alpha1.CommitStatusPhase
+}
+
+// HTTPEXecutor performs the WebRequestCommitStatus HTTP round-trip for the given template data
+// (render, auth, transport). The simulator supplies an implementation that returns a mock response
+// or ErrHTTPResponseRequiredWhenTriggerFires when none is set.
+type HTTPEXecutor interface {
+	Execute(ctx context.Context, wrcs *promoterv1alpha1.WebRequestCommitStatus, td TemplateData) (HTTPResponse, error)
+}
+
+// triggerExprData builds the expression data map for trigger and trigger output expressions.
+func (td TemplateData) triggerExprData() map[string]any {
+	return map[string]any{
+		"Branch":                 td.Branch,
+		"Phase":                  td.Phase,
+		"PromotionStrategy":      td.PromotionStrategy,
+		"WebRequestCommitStatus": td.WebRequestCommitStatus,
+		"TriggerOutput":          td.TriggerOutput,
+		"ResponseOutput":         td.ResponseOutput,
+		"SuccessOutput":          td.SuccessOutput,
+	}
+}
+
+// successWhenExprData builds the expression data map for success.when expressions.
+// It mirrors triggerExprData and adds Response: the HTTP response map when a request was
+// made this reconcile, or nil otherwise.
+func successWhenExprData(td TemplateData, resp *HTTPResponse) map[string]any {
+	exprData := td.triggerExprData()
+	if resp != nil {
+		exprData["Response"] = map[string]any{
+			"StatusCode": resp.StatusCode,
+			"Body":       resp.Body,
+			"Headers":    resp.Headers,
+		}
+	} else {
+		exprData["Response"] = nil
+	}
+	return exprData
+}
+
+// withLatestOutputs returns a copy of the template data with ResponseOutput, TriggerOutput, and SuccessOutput
+// updated from the latest HTTP response, trigger evaluation, and success evaluation. Used before upserting
+// CommitStatuses so description/URL templates reflect current data.
+func (td TemplateData) withLatestOutputs(responseDataJSON *apiextensionsv1.JSON, newTriggerData map[string]any, successDataJSON *apiextensionsv1.JSON) TemplateData {
+	result := td
+	if responseDataJSON != nil {
+		if data, err := unmarshalJSONMap(responseDataJSON); err == nil && data != nil {
+			result.ResponseOutput = data
+		}
+	}
+	if newTriggerData != nil {
+		result.TriggerOutput = newTriggerData
+	}
+	if successDataJSON != nil {
+		if data, err := unmarshalJSONMap(successDataJSON); err == nil && data != nil {
+			result.SuccessOutput = data
+		}
+	}
+	return result
+}
+
+// unmarshalJSONMap unmarshals an apiextensionsv1.JSON into a map. Returns (nil, nil) when raw is nil.
+func unmarshalJSONMap(raw *apiextensionsv1.JSON) (map[string]any, error) {
+	if raw == nil {
+		return nil, nil
+	}
+	result := make(map[string]any)
+	if err := json.Unmarshal(raw.Raw, &result); err != nil {
+		return nil, fmt.Errorf("failed to unmarshal JSON map: %w", err)
+	}
+	return result, nil
+}
+
+// marshalJSONMap marshals a map into an apiextensionsv1.JSON. Returns (nil, nil) when data is nil.
+func marshalJSONMap(data map[string]any) (*apiextensionsv1.JSON, error) {
+	if data == nil {
+		return nil, nil
+	}
+	raw, err := json.Marshal(data)
+	if err != nil {
+		return nil, fmt.Errorf("failed to marshal JSON map: %w", err)
+	}
+	return &apiextensionsv1.JSON{Raw: raw}, nil
 }
